Daily Commit
This is the first major commit, at the end of day 1 of development.

This commit includes:

Building the host builder, to allow connection to Discord.

Building of CommandHandler Service to facilitate creation of commands for the bot.

Creates the DBContext and DBContextFactory for the project.

Creates the DAL for the project.

Creates Various Models for the project, to be used for various modules throughout.

Added Custom Embed Builder to be used later when integrating modules.

Various minor things, such as icons, colours etc.