Memes
This commit adds a new meme command, which allows users to query an API endpoint and return a meme from a subreddit.

This commit also adds a Serializer and Convertor for the API endpoint making it easier to parse.

We also added a HTTPClientFactory to facilitate querying these endpoints in a compliant way.